requirements.txt
# Dependencias de Python para el proyecto Titanic ML

# Análisis de datos
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2

# Visualización
matplotlib==3.8.2
seaborn==0.13.0

# Guardar/cargar modelos
joblib==1.3.2

# Django y Django REST Framework
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Utilidades
python-dotenv==1.0.0
